webapi/webapi-webrtc-w3c-tests/webrtc/blink/RTCPeerConnection-createOffer.html

<!DOCTYPE html>
<!--
Test converted from blink:
https://github.com/crosswalk-project/blink-crosswalk/blob/master/LayoutTests/fast/mediastream/RTCPeerConnection-createOffer.html
-->

<meta charset=utf-8>
<title>WebRTC Test: RTCPeerConnection createOffer</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel=help href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#widl-RTCPeerConnection-createOffer-void-RTCSessionDescriptionCallback-successCallback-RTCPeerConnectionErrorCallback-failureCallback-MediaConstraints-constraints">
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<div id=log></div>
<script>

var t = async_test(document.title);

var pc;
if (typeof RTCPeerConnection != "undefined") {
  pc = new RTCPeerConnection(null, null);
} else {
  pc = new webkitRTCPeerConnection(null, null);
}

function requestFailed2(reason) {
  errorReason = reason;
  t.step(function () {
    assert_equals(errorReason, "TEST_ERROR");
  });
  t.done();
}

function requestFailed1() {
  t.step(function () {
    assert_unreached("requestFailed1 was called");
  });
  t.done();
}

function requestSucceeded1(sd) {
  sessionDescription = sd;
  t.step(function () {
    assert_equals(sessionDescription.type, "offer");
  });
  pc.createOffer(t.step_func(function () {
    t.done();
  }), requestFailed2);
}

t.step(function () {
  pc.createOffer(requestSucceeded1, requestFailed1);
});

</script>
